improve `NequIP` (#16)
The original `NequIP` requires that you know in advance all the elements
you wish to train on, in all pre-training/fine-tuning stages. This is
due to the `OneHot` nature of the atom-type embedding used in the
self-interaction branch of the node update function.

This PR abstracts the nature of this atom-type embedding out of the
`_BaseNequIP` base class, and provides two concrete implementations:
`OneHotNequIP` (the original) and `ZEmbeddingNequIP`, where the
atom-type embedding is a learnable, dense embedding of the atomic
number. This side-steps the requirement to know _a priori_ how many
elements you will be training on in future fine-tuning steps.